+++ R/create_kegg_db.R
@@ -3,14 +3,16 @@
 ##' 
 ##' @title create_kegg_db
 ##' @param species one of KEGG supported species, e.g. hsa for human
+##' @param author a string for the Author field in the generated package DESCRIPTION
+##' @param maintainer a string for the Maintainer field in the generated package DESCRIPTION
 ##' @return KEGG.db package generated in working directory
 ##' @export
 ##' @author Guangchuang Yu and Ziru Chen
-create_kegg_db <- function(species) {
+create_kegg_db <- function(species, author = NULL, maintainer = NULL) {
     packagedir <- tempfile() # tempdir() maynot empty
 
     ## skeleton
-    prepare_pkg_skeleton(packagedir)
+    prepare_pkg_skeleton(packagedir, author = author, maintainer = maintainer)
 
     ## sqlite
     sqlite_path <- paste(packagedir, "inst", "extdata", sep=.Platform$file.sep)
@@ -21,7 +23,7 @@ create_kegg_db <- function(species) {
 }
 
 
-prepare_pkg_skeleton <- function(packagedir) {
+prepare_pkg_skeleton <- function(packagedir, author = NULL, maintainer = NULL) {
     .fcp <- function(..., todir, file) {
         file.copy(from = system.file("KEGG.db", ..., file, package = "createKEGGdb"),
                   to = paste(todir, file, sep = .Platform$file.sep))
@@ -46,49 +48,88 @@ prepare_pkg_skeleton <- function(packagedir) {
     .fcp(todir = packagedir, file = "DESCRIPTION")
     .fcp(todir = packagedir, file = "LICENSE")
     .fcp(todir = packagedir, file = "NAMESPACE")
+
+    if (!is.null(author) || !is.null(maintainer)) {
+        desc_path <- file.path(packagedir, "DESCRIPTION")
+        desc_lines <- readLines(desc_path, warn = FALSE)
+        if (!is.null(author)) {
+            desc_lines <- sub("^Author:.*$", paste0("Author: ", author), desc_lines)
+        }
+        if (!is.null(maintainer)) {
+            desc_lines <- sub("^Maintainer:.*$", paste0("Maintainer: ", maintainer), desc_lines)
+        }
+        writeLines(desc_lines, con = desc_path)
+    }
 }
 
 
 ##' @importFrom magrittr %<>%
 get_path2name <- function(species){
+  safe_kegg_list <- function(db, organism = NULL) {
+    tryCatch(
+      {
+        if (is.null(organism)) {
+          clusterProfiler:::kegg_list(db)
+        } else {
+          clusterProfiler:::kegg_list(db, organism)
+        }
+      },
+      error = function(e) NULL
+    )
+  }
+
   if (length(species) == 1) {
-    keggpathid2name.df <- clusterProfiler:::kegg_list("pathway", species)
+    keggpathid2name.df <- safe_kegg_list("pathway", species)
   } else {
     keggpathid2name.list <- vector("list", length(species))
     names(keggpathid2name.list) <- species
     for (i in species) {
-      keggpathid2name.list[[i]] <- clusterProfiler:::kegg_list("pathway", i)    
+      keggpathid2name.list[[i]] <- safe_kegg_list("pathway", i)
     }
-    keggpathid2name.df <- do.call(rbind, keggpathid2name.list)
-    rownames(keggpathid2name.df) <- NULL
+    keggpathid2name.list <- keggpathid2name.list[!vapply(keggpathid2name.list, is.null, logical(1))]
+    if (length(keggpathid2name.list) == 0) {
+      keggpathid2name.df <- NULL
+    } else {
+      keggpathid2name.df <- do.call(rbind, keggpathid2name.list)
+      rownames(keggpathid2name.df) <- NULL
+    }
   }
-  keggpathid2name.df[,2] <- sub("\\s-\\s[a-zA-Z ]+\\(\\w+\\)$", "", keggpathid2name.df[,2])
-  # keggpathid2name.df[,1] %<>% gsub("path:map", "", .)
-  colnames(keggpathid2name.df) <- c("path_id","path_name")
-  return(keggpathid2name.df)
+
+  if (is.null(keggpathid2name.df) || nrow(keggpathid2name.df) == 0) {
+    data.frame(path_id = character(0), path_name = character(0), stringsAsFactors = FALSE)
+  } else {
+    keggpathid2name.df[,2] <- sub("\\s-\\s[a-zA-Z ]+\\(\\w+\\)$", "", keggpathid2name.df[,2])
+    colnames(keggpathid2name.df) <- c("path_id","path_name")
+    keggpathid2name.df
+  }
 }
 
 
 ##' @importFrom magrittr %<>%
 download.organisms.KEGG <- function(organism) {
-  keggpathid2extid.df <- clusterProfiler:::kegg_link(organism, "pathway")
-  if (is.null(keggpathid2extid.df)){
-    write(paste(Sys.time(),"Pathway data of",organism,"is null."), stderr())
-  }else{
-    message(paste0(Sys.time()," Getting KEGG data of ",organism,"."))
-    keggpathid2extid.df[,1] %<>% gsub("[^:]+:", "", .)
-    keggpathid2extid.df[,2] %<>% gsub("[^:]+:", "", .)
-    colnames(keggpathid2extid.df) <- c("pathway_id","gene_or_orf_id")
-    message(paste(Sys.time(),"KEGG data of",organism,"has been downloaded."))
+  keggpathid2extid.df <- tryCatch(
+    clusterProfiler:::kegg_link(organism, "pathway"),
+    error = function(e) NULL
+  )
+
+  if (is.null(keggpathid2extid.df) || nrow(keggpathid2extid.df) == 0) {
+    write(paste(Sys.time(), "Pathway data of", organism, "is null."), stderr())
+    return(NULL)
   }
-  return(keggpathid2extid.df)
+
+  message(paste0(Sys.time(), " Getting KEGG data of ", organism, "."))
+  keggpathid2extid.df[,1] %<>% gsub("[^:]+:", "", .)
+  keggpathid2extid.df[,2] %<>% gsub("[^:]+:", "", .)
+  colnames(keggpathid2extid.df) <- c("pathway_id","gene_or_orf_id")
+  message(paste(Sys.time(), "KEGG data of", organism, "has been downloaded."))
+  keggpathid2extid.df
 }
 
 
 get_organisms_list <- function(db){
-  organisms <- clusterProfiler:::kegg_list(db)
-  organisms_list <- as.character(organisms[,2])
-  return(organisms_list)
+  organisms <- tryCatch(clusterProfiler:::kegg_list(db), error = function(e) NULL)
+  if (is.null(organisms) || nrow(organisms) == 0) return(character(0))
+  as.character(organisms[,2])
 }
 
 
@@ -106,27 +147,63 @@ prepare_kegg_db <- function(organisms, sqlite_path) {
   drv <- dbDriver("SQLite")
   db <- dbConnect(drv, dbname=dbfile)
 
-  KEGGPATHID2NAME <- get_path2name(organisms)
-  ###################################################
-  ### put the pathway2name data into the tables 
-  ###################################################
-  dbWriteTable(conn = db, "pathway2name", KEGGPATHID2NAME, row.names=FALSE)
-  
-  if (length(organisms) == 1){
-    if(organisms == "all"){
-      organisms <- get_organisms_list("organism")
-    }
+  if (length(organisms) >= 1 && any(organisms == "all")) {
+    organisms <- get_organisms_list("organism")
   }
-  for(organism in organisms){
+
+  use_map_pathway_names <- length(organisms) > 50
+
+  dbWriteTable(
+    conn = db,
+    name = "pathway2gene",
+    value = data.frame(pathway_id = character(0), gene_or_orf_id = character(0), stringsAsFactors = FALSE),
+    row.names = FALSE
+  )
+
+  pathway2gene_count <- 0L
+  pathway2name_count <- 0L
+
+  map_names <- NULL
+  if (use_map_pathway_names) {
+    dbWriteTable(
+      conn = db,
+      name = "pathway2name",
+      value = data.frame(path_id = character(0), path_name = character(0), stringsAsFactors = FALSE),
+      row.names = FALSE
+    )
+
+    kegg_map_pathway <- tryCatch(clusterProfiler:::kegg_list("pathway"), error = function(e) NULL)
+    if (!is.null(kegg_map_pathway) && nrow(kegg_map_pathway) > 0) {
+      kegg_map_pathway[, 2] <- sub("\\s-\\s[a-zA-Z ]+\\(\\w+\\)$", "", kegg_map_pathway[, 2])
+      map_ids <- gsub("[^:]+:", "", kegg_map_pathway[, 1])
+      map_digits <- sub("^[^0-9]+", "", map_ids)
+      map_names <- as.character(kegg_map_pathway[, 2])
+      names(map_names) <- map_digits
+    }
+  } else {
+    KEGGPATHID2NAME <- get_path2name(organisms)
+    dbWriteTable(conn = db, "pathway2name", KEGGPATHID2NAME, row.names = FALSE)
+    pathway2name_count <- nrow(KEGGPATHID2NAME)
+  }
+
+  for (organism in organisms) {
     KEGGPATHID2EXTID <- download.organisms.KEGG(organism)
-    if(!is.null(KEGGPATHID2EXTID)){
-      ###################################################
-      ### put the pathway2gene data into the tables 
-      ###################################################
-      # 数据是直接添加进去，不会自动去重
-      dbWriteTable(conn = db, "pathway2gene", KEGGPATHID2EXTID, row.names=FALSE,append = TRUE)
-      message(paste(Sys.time(),"KEGG data of",organism,"has been added to the sqlite database."))
-      
+    if (!is.null(KEGGPATHID2EXTID)) {
+      dbWriteTable(conn = db, "pathway2gene", KEGGPATHID2EXTID, row.names = FALSE, append = TRUE)
+      pathway2gene_count <- pathway2gene_count + nrow(KEGGPATHID2EXTID)
+      message(paste(Sys.time(), "KEGG data of", organism, "has been added to the sqlite database."))
+
+      if (use_map_pathway_names && !is.null(map_names)) {
+        path_ids <- unique(KEGGPATHID2EXTID[, 1])
+        path_digits <- sub("^[^0-9]+", "", path_ids)
+        KEGGPATHID2NAME_org <- data.frame(
+          path_id = path_ids,
+          path_name = unname(map_names[path_digits]),
+          stringsAsFactors = FALSE
+        )
+        dbWriteTable(conn = db, "pathway2name", KEGGPATHID2NAME_org, row.names = FALSE, append = TRUE)
+        pathway2name_count <- pathway2name_count + nrow(KEGGPATHID2NAME_org)
+      }
     }
   }
   
@@ -150,8 +227,8 @@ prepare_kegg_db <- function(organisms, sqlite_path) {
   colnames(metadata) <- c("name", "value") #makeAnnDbPkg规定的
   dbWriteTable(conn = db, "metadata", metadata, row.names=FALSE)
   
-  map.counts <- rbind(c("pathway2name", nrow(KEGGPATHID2NAME)),
-                      c("pathway2gene", nrow(KEGGPATHID2EXTID)))
+  map.counts <- rbind(c("pathway2name", pathway2name_count),
+                      c("pathway2gene", pathway2gene_count))
   map.counts <- as.data.frame(map.counts)
   colnames(map.counts) <- c("map_name","count")
   dbWriteTable(conn = db, "map_counts", map.counts, row.names=FALSE)