+++ NEWS.md
@@ -1,4 +1,9 @@
 
+# createKEGGdb 0.0.4
+
++ fix `species="all"` and missing pathway species handling (2026-1-29, Thu)
++ allow setting Author/Maintainer for generated KEGG.db package (2026-1-29, Thu)
+
 
 # createKEGGdb 0.0.3